openssl_trace.c: declare bytes_count as int, because open ssl returns as int, not as long. Prevent aliasing of -1 to 2^32-1 e.g., prevent bug where we continue to process data that we should skip.
Summary: openssl_trace.c: declare bytes_count as int, because open ssl returns as int, not as long. Prevent aliasing of -1 to 2^32-1 e.g., prevent bug where we continue to process data that we should skip.

Test Plan: Jenkins

Reviewers: yzhao, oazizi, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PP-2393

Differential Revision: https://phab.corp.pixielabs.ai/D7152

GitOrigin-RevId: ad8029a9c87587ab961ec198cd090eb324f527ab

+++ src/stirling/source_connectors/socket_tracer/bcc_bpf/openssl_trace.c
@@ -7,7 +7,7 @@
 static __inline void process_openssl_data(struct pt_regs* ctx, uint64_t id,
                                           const enum TrafficDirection direction,
                                           const struct data_args_t* args) {
-  ssize_t bytes_count = PT_REGS_RC(ctx);
+  int bytes_count = PT_REGS_RC(ctx);
   process_data(/* vecs */ false, ctx, id, direction, args, bytes_count, /* ssl */ true);
 }
